musicplatform/requirements.txt
django~=4.1.2
django-model-utils~=4.2.0
djangorestframework~=3.14.0
django-rest-knox~=4.2.0
